Create Arrays DS
Given an array, A, of N integers, print each element in reverse order as a single line of space-separated integers.
The first line contains an integer,  N(the number of integers in A). 
The second line contains N space-separated integers describing A.
Print all N integers in A in reverse order as a single line of space-separated integers.